Fix Request Duration Metric (#59)
The request duration metric was not recorded correctly. This is now
fixed, by splitting the instrumentation for metrics, logs and traces, so
that the metric and logs are written once the request is completed,
while traces are handled separately.

Within the traces instrumentation we now also handle the recovering from
panics, so that the additional recoverer middleware is no longer needed
and has been removed.

Last but not least the `httpsnoop` package has be updated to the latest
commit, which includes some important fixes.