Fix Missing Trace ID in Logs
The request logs didn't contain the trace id, after refactoring the
instrumentation in #59. This is now fixed, by adding the trace id to the
request information in the tracing middleware and checking the request
info next to the trace context in the logger.

This commit also adds the span id as an additional field in the logs.